requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
greenlet==2.0.2
psycopg2-binary==2.9.7
pytz==2023.3
SQLAlchemy==2.0.20
sqlparse==0.4.4
typing_extensions==4.7.1
